+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Add order.wf.edit_requirements event (ericof).
 * Add CustomerUserProfile and BriefyUserProfile events (ericof).
 * Add Slack actions to CustomerUserProfile and BriefyUserProfile (ericof).
+* Add Slack actions to Professional (ericof).
 
 2.0.1 (2017-02-14)
 ------------------